requirements.txt
Django==2.2
django-role-permissions==2.2.1
python-dotenv==0.10.3
